+++ tools/perf/tests/bp_signal.c
@@ -11,6 +11,7 @@
  */
 #define __SANE_USERSPACE_TYPES__
 
+#include <inttypes.h>
 #include <stdlib.h>
 #include <stdio.h>
 #include <unistd.h>
@@ -111,7 +112,7 @@ static int __event(bool is_x, void *addr, int sig)
 
 	pe.config = 0;
 	pe.bp_type = is_x ? HW_BREAKPOINT_X : HW_BREAKPOINT_W;
-	pe.bp_addr = (unsigned long) addr;
+	pe.bp_addr = (unsigned long)(uintptr_t) addr;
 	pe.bp_len = is_x ? default_breakpoint_len() : sizeof(long);
 
 	pe.sample_period = 1;
