+++ tools/perf/tests/mmap-thread-lookup.c
@@ -196,7 +196,7 @@ static int mmap_events(synth_cb synth)
 		pr_debug("looking for map %p\n", td->map);
 
 		thread__find_map(thread, PERF_RECORD_MISC_USER,
-				 (unsigned long) (td->map + 1), &al);
+				 (unsigned long)(uintptr_t) (td->map + 1), &al);
 
 		thread__put(thread);
